get user menu working
