+++ app/app/(app)/(tabs)/ride/_layout.tsx
@@ -1,4 +1,6 @@
+import { getNavigationMenuFromOptions } from "@/components/Menu.utils";
 import { RideMenu } from "@/components/RideMenu";
+import { UserMenu, useUserMenuOptions } from "@/components/UserMenu";
 import { isAndroid, isWeb } from "@/utils/constants";
 import { Stack } from "expo-router"
 
@@ -14,6 +16,22 @@ export default function Layout() {
       />
       <Stack.Screen name="map" options={{ headerTitle: "Beeper Map" }} />
       <Stack.Screen name="pick" options={{ headerTitle: "Choose Beeper" }} />
+       <Stack.Screen
+        options={(route) => {
+          const params = route.route.params as { id: string };
+          return {
+            headerRight: () => {
+              return <UserMenu userId={params.id} />;
+            },
+            unstable_headerRightItems: () => {
+              const options = useUserMenuOptions(params.id);
+              return getNavigationMenuFromOptions(options);
+            },
+            headerTitle: "User"
+          }
+        }}
+        name="user/[id]/index"
+      />
     </Stack>
   );
 }
